% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_cells_lt}
\alias{add_cells_lt}
\title{Add cells while bypassing checks}
\usage{
add_cells_lt(obj, n_cells, matrix.data, layers.data)
}
\arguments{
\item{obj}{The \code{loom} object to which cells are to be added.}

\item{n_cells}{Number of cells to be added; the dimensions of the data are assumed to be
compatible with this number.}

\item{matrix.data}{A Matrix with \code{n_cells} rows and as many columns as there're genes;
data for new cells to be added to the \code{matrix} dataset of the \code{loom} object.}

\item{layers.data}{A named list of matrices with the same dimension as \code{matrix.data}
for layer data for the new cells.}
}
\description{
The \code{add.cells} method of \code{loom} class does a lot of checks on the dimensions of
data for cells to be added to a \code{loom} object. However, here, while I do need to add cells,
I already know the dimensions and that they work, so I can bypass the checks and save some time,
which can be a lot if there are many iterations.
Here I assume that the matrices will not be transposed. This function is basically a lightweight
version of the method \code{loom}.
}
