% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_col_attrs_lt}
\alias{add_col_attrs_lt}
\title{Add column attributes for all the added cells}
\usage{
add_col_attrs_lt(obj, n_cells, attributes.data)
}
\arguments{
\item{obj}{The \code{loom} object to which cells are to be added.}

\item{n_cells}{Number of cells to be added; the dimensions of the data are assumed to be
compatible with this number.}

\item{attributes.data}{A named list of matrices (with \code{n_cells} columns) or vectors
(with length \code{n_cells}) for column attributes for the new cells.}
}
\description{
In case there are many column attributes, and many chunks of new cells to be added, iterating
over all those attributes for each chunk would be very inefficient. This function adds the
column attributes to the new cells after they have all been added to the \code{loom} object.
}
