% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimodality_coefficient.R
\name{bimodality_coefficient}
\alias{bimodality_coefficient}
\title{bimodality_coefficient}
\usage{
bimodality_coefficient(x)
}
\arguments{
\item{x}{Guassian kernel density estimation representing pANN distribution}
}
\value{
Bimodality coefficient value
}
\description{
Internal function to compute bimodality coefficient during BCmvn computation
and pK estimation.
}
\references{
Taken from the 'modes' R package (v0.7)
}
\author{
Chris McGinnis
}
