% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extend_col_attrs}
\alias{extend_col_attrs}
\title{Extend column attributes when adding cells to loom object}
\usage{
extend_col_attrs(n_cells, obj, cell.names = "cell_names")
}
\arguments{
\item{n_cells}{Number of cells to add.}

\item{obj}{The loom object to which cells are to be added. Only column attributes with 1
or 2 dimensions are supported.}

\item{cell.names}{Name of dataset in loom object that stores cell names}
}
\value{
A data frame with dummy attributes of the right column names, right
data types, and the right dimensions.
}
\description{
The purpose of this function is to make sure that the doublet loom object
can be properly combined with the copy of the real cell loom object.
In oroder to combine properly, the two loom objects must have the same column and row
attributes and the same layers. This function takes care of the column attributes by taking
\code{n_cells} rows from column attributes of the loom object with real cells, so the data
type will match. I can simply copy the row attribute (for genes) from the object with real cells
to that with doublets.
}
