% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_all_attributes}
\alias{get_all_attributes}
\title{Get all column/row attributes}
\usage{
get_all_attributes(obj, MARGIN)
}
\arguments{
\item{obj}{The loom object whose attributes will be retrieved. Only 1 or 2 dimensions are
supported.}

\item{MARGIN}{Margin; 1 means row (genes), and 2 means column (cells)}
}
\value{
A list with all attributes of the margin
}
\description{
The method get.attributes.df only returns attributes with one dimension. This function
will return a list with attributes with both 1 and 2 dimensions.
}
