% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurtosis.R
\name{kurtosis}
\alias{kurtosis}
\title{kurtosis}
\usage{
kurtosis(x)
}
\arguments{
\item{x}{Guassian kernel density estimation representing pANN distribution}
}
\value{
Kurtosis value
}
\description{
Internal function to compute bimodality coefficient during BCmvn computation
and pK estimation.
}
\references{
Taken from the 'modes' R package (v0.7).
}
\author{
Chris McGinnis
}
