% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_small}
\alias{pbmc_small}
\title{A SeuratObject containing 230 features across 80 cells within 1 assay}
\format{
A SeuratObject containing 230 features across 80 cells within 1 assay
}
\source{
From the SeuratObject package: https://github.com/satijalab/seurat-object/blob/main/data/pbmc_small.rda
}
\usage{
data(pbmc_small)
}
\description{
A SeuratObject containing 230 features across 80 cells within 1 assay for
a subset human PBMC single cell RNA-seq experiment.
}
\keyword{datasets}
