% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness.R
\name{skewness}
\alias{skewness}
\title{skewness}
\usage{
skewness(x)
}
\arguments{
\item{x}{Guassian kernel density estimation representing pANN distribution}
}
\value{
Skewness value
}
\description{
Internal function to compute skewness during BCmvn computation and pK
estimation.
}
\examples{

## Internal to bimodality_coefficient
G <- skewness(x)

}
\references{
Taken from the 'modes' R package (v0.7).
}
\author{
Chris McGinnis
}
