% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeSweep.R
\name{summarizeSweep}
\alias{summarizeSweep}
\title{summarizeSweep}
\usage{
summarizeSweep(sweep.list, GT = FALSE, GT.calls = NULL)
}
\arguments{
\item{sweep.list}{List of pANN vectors across pN-pK space, as produced by
doubletFinder_ParamSweep.}

\item{GT}{Logical set to TRUE when ground-truth doublet classifications are
available for ROC analysis. Default set to FALSE.}

\item{GT.calls}{An nCell-length character vector of ground-truth doublet
classifications (e.g., "Singlet" or "Doublet") used to gauge performance of
logistic regression models trained using pANN vectors during ROC analysis.}
}
\value{
Dataframe with bimodality coefficient values at each pN-pK parameter
set. If GT = TRUE, dataframe also includes AUC for each pN-pK parameter set
computed during ROC analysis.
}
\description{
Summarizes results from doubletFinder_ParamSweep, computing the bimodality
coefficient across pN and pK parameter space. If ground-truth doublet
classifications are available, then ROC analysis is performed, enabling
optimal DoubletFinder parameter selection.
}
\examples{
data(pbmc_small)
seu <- pbmc_small
sweep.list <- paramSweep(seu)
sweep.stats <- summarizeSweep(sweep.list, GT = FALSE)
bcmvn <- find.pK(sweep.stats)

}
\author{
Chris McGinnis
}
